refactor: extracted WorkoutBuilder's partial widgets into own files
