+++ lib/layouts/home_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:just_another_workout_timer/layouts/workout_builder/workout_builder.dart';
 import 'package:just_another_workout_timer/layouts/workout_runner.dart';
 
 import '../generated/l10n.dart';
@@ -7,7 +8,6 @@ import 'settings_page.dart';
 import '../utils/storage_helper.dart';
 import '../utils/utils.dart';
 import '../utils/workout.dart';
-import 'workout_builder.dart';
 
 /// Main screen
 class HomePage extends StatefulWidget {
